+++ src/jcds/charts/__init__.py
@@ -3,7 +3,17 @@ import jcds.utils
 
 # Import your functions from internal files
 from .roc import plot_roc
+from .boxplots import outlier_boxplots
+from .missing import missing_data_heatmap
+from .categorical import categorical_barplot
+from .correlation import correlation_heatmap
 
 help = jcds.utils._make_module_help(sys.modules[__name__])
 
-__all__ = ["plot_roc", "help"]
+__all__ = ["plot_roc", 
+           "outlier_boxplots", 
+           "missing_data_heatmap", 
+           "correlation_heatmap", 
+           "categorical_barplot", 
+           "help"
+        ]